backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-docx==1.1.0
pyngrok==7.0.0

